Performance tuning

+++ app/src/main/java/com/kotlin/practice/todoapp/contents/models/TaskData.kt
@@ -1,7 +1,7 @@
 package com.kotlin.practice.todoapp.contents.models
 
 data class TaskData(
-    val primaryKey: Int,
+    val primaryKey: Long,
     var task: String,
     var isChecked: Boolean
 )
